% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set of x and y numeric vectors}
\format{
## `ddt`
A data frame with 30 rows and 2 columns:
\describe{
  \item{x}{numeric vector}
  \item{y}{numeric vector}
  \item{double}{double}
  ...
}
}
\source{
<Installment 4 assignment example input>
}
\usage{
ddt
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
