% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inst4Rox.R
\name{tConstr}
\alias{tConstr}
\title{Title t-test Installment 4}
\usage{
tConstr(x, y, alpha, independentSamp = TRUE, equalVar = TRUE)
}
\arguments{
\item{x}{X Vector (numeric)}

\item{y}{Y vector (numeric)}

\item{alpha}{alpha/error value (numeric)}

\item{independentSamp}{Is independent? (default TRUE)}

\item{equalVar}{Has equal variance? (default TRUE)}
}
\value{
t-test and input results
}
\description{
Title t-test Installment 4
}
\examples{
\dontrun{tConstr(x=rnorm(30,5,2), y=rnorm(30,3,2), alpha=0.05)}
}
